% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{length,sparse_numeric-method}
\alias{length,sparse_numeric-method}
\title{Length of a sparse_numeric vector}
\usage{
\S4method{length}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}
}
\value{
Integer length of the underlying full vector.
}
\description{
Length of a sparse_numeric vector
}
