% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\alias{mean}
\title{mean of sparse_numeric}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{...}{Additional arguments (ignored)}
}
\value{
A numeric scalar
}
\description{
mean of sparse_numeric
}
