% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Compute a vector norm}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{Object to compute the norm of.}

\item{...}{Ignored.}
}
\value{
A numeric scalar.
}
\description{
Generic for computing norms of objects. For a
\linkS4class{sparse_numeric} vector this returns the Euclidean (L2)
norm \eqn{sqrt(sum(x_i^2))}.
}
