% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot two sparse_numeric vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{\code{sparse_numeric} vectors of the same length.}

\item{...}{Additional arguments (ignored).}
}
\description{
Plot two sparse_numeric vectors
}
