% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_arith_ops}
\alias{sparse_arith_ops}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Arithmetic operators for sparse_numeric vectors}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1, e2}{\code{sparse_numeric} vectors.}
}
\description{
Elementwise addition, subtraction, and multiplication using the
standard \code{+}, \code{-}, and \code{*} operators.
}
