% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric}
\alias{sparse_numeric}
\title{Construct a sparse_numeric vector from a numeric vector}
\usage{
sparse_numeric(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An object of class \code{sparse_numeric}.
}
\description{
Construct a sparse_numeric vector from a numeric vector
}
