% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\docType{methods}
\name{sparse_ops}
\alias{sparse_ops}
\alias{sparse_add}
\alias{sparse_sub}
\alias{sparse_mult}
\alias{sparse_crossprod}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Sparse arithmetic operations}
\usage{
sparse_add(x, y, ...)

sparse_sub(x, y, ...)

sparse_mult(x, y, ...)

sparse_crossprod(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x, y}{\linkS4class{sparse_numeric} objects.}

\item{...}{Ignored.}

\item{e1, e2}{\linkS4class{sparse_numeric} objects for the arithmetic operators (\code{+}, \code{-}, \code{*}).}
}
\value{
\itemize{
\item \code{sparse_add()}, \code{sparse_sub()}, and \code{sparse_mult()} return a
\linkS4class{sparse_numeric} object.
\item \code{sparse_crossprod()} returns a numeric scalar (dot product).
}
}
\description{
Generics and methods implementing basic arithmetic on
\linkS4class{sparse_numeric} vectors.
}
