% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse_numeric vectors}
\usage{
sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{y}{A \code{sparse_numeric} vector.}

\item{...}{Ignored; included for method compatibility.}
}
\value{
A \code{sparse_numeric} vector representing \code{x - y}.
}
\description{
Elementwise subtraction of two sparse numeric vectors of the same length.
}
