% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sparseNumeric}
\alias{sparseNumeric-package}
\alias{sparseNumeric}
\title{sparseNumeric: sparse numeric vector class and methods}
\description{
Implements an S4 sparse numeric vector class with methods for arithmetic, plotting, computing means and norms, and standardizing vectors without materializing dense representations unless necessary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/masoncancode/sparseNumeric}
  \item \url{https://masoncancode.github.io/sparseNumeric/}
  \item Report bugs at \url{https://github.com/masoncancode/sparseNumeric/issues}
}

}
\author{
\strong{Maintainer}: Mason Reed \email{you@example.com}

}
