% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sum,sparse_numeric-method}
\alias{sum,sparse_numeric-method}
\title{Sum of a sparse_numeric vector}
\usage{
\S4method{sum}{sparse_numeric}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored.}

\item{na.rm}{Logical; if TRUE, remove NA values in the stored entries.}
}
\value{
Numeric scalar giving the sum of all elements (zeros included).
}
\description{
Sum of a sparse_numeric vector
}
