% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{summary,sparse_numeric-method}
\alias{summary,sparse_numeric-method}
\title{Summary for sparse_numeric}
\usage{
\S4method{summary}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse_numeric vector}
}
\value{
Prints summary statistics to console
}
\description{
Summary for sparse_numeric
}
