% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{length,sparse_numeric-method}
\alias{length,sparse_numeric-method}
\title{Length method for sparse_numeric}
\usage{
\S4method{length}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object.}
}
\value{
The length of the underlying full vector.
}
\description{
Length method for sparse_numeric
}
