% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseNumeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric vector (including zeros)}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{sparse_numeric}

\item{...}{unused}
}
\value{
numeric scalar
}
\description{
Mean of a sparse_numeric vector (including zeros)
}
