% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \linkS4class{sparse_numeric} object.}

\item{...}{Ignored.}
}
\value{
Numeric scalar mean.
}
\description{
Computes the mean of the underlying full vector (including implicit
zeros) without coercing to a dense representation.
}
