% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseNumeric.R
\name{norm}
\alias{norm}
\title{Euclidean norm of a sparse_numeric vector}
\usage{
norm(x, ...)
}
\arguments{
\item{x}{sparse_numeric}

\item{...}{unused}
}
\value{
numeric scalar
}
\description{
Euclidean norm of a sparse_numeric vector
}
