% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm,sparse_numeric,ANY-method}
\alias{norm,sparse_numeric,ANY-method}
\alias{norm}
\title{Norm of sparse_numeric}
\usage{
\S4method{norm}{sparse_numeric,ANY}(x, type, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{type}{Norm type}

\item{...}{Additional arguments}
}
\value{
A numeric scalar
}
\description{
Norm of sparse_numeric
}
