% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Norm (L2) generic}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric vector.}

\item{...}{Ignored, included for consistency with the generic.}
}
\value{
A numeric scalar containing the Euclidean norm.
}
\description{
Norm (L2) generic

L2 norm of a sparse_numeric vector
}
