% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot two sparse_numeric vectors for comparison}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{y}{A sparse_numeric vector}

\item{...}{Additional arguments passed to plot()}
}
\description{
Plot two sparse_numeric vectors for comparison
}
