% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseNumeric.R
\name{sparse_add}
\alias{sparse_add}
\title{Add two sparse_numeric vectors}
\usage{
sparse_add(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric objects of same length}

\item{...}{unused}
}
\value{
sparse_numeric
}
\description{
Add two sparse_numeric vectors
}
