% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Sparse Vector Addition}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{y}{A sparse_numeric vector}

\item{...}{Additional arguments}
}
\description{
Sparse Vector Addition

Add two sparse_numeric vectors
}
