% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{sparse_add generic}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric vectors.}

\item{...}{Ignored, included for consistency with the generic.}
}
\value{
A sparse_numeric vector representing \code{x + y}.
}
\description{
sparse_add generic
}
