% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseNumeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Dot product of two sparse_numeric vectors}
\usage{
sparse_crossprod(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric objects of same length}

\item{...}{unused}
}
\value{
numeric scalar
}
\description{
Dot product of two sparse_numeric vectors
}
