% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Sparse Vector Crossproduct}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{y}{A sparse_numeric vector}

\item{...}{Additional arguments}
}
\description{
Sparse Vector Crossproduct

Cross-product (inner product) of two sparse_numeric vectors
}
