% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{sparse_crossprod generic}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric vectors.}

\item{...}{Ignored, included for consistency with the generic.}
}
\value{
A numeric scalar equal to the dot product of \code{x} and \code{y}.
}
\description{
sparse_crossprod generic
}
