% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseNumeric.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Multiply two sparse_numeric vectors elementwise}
\usage{
sparse_mult(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric objects of same length}

\item{...}{unused}
}
\value{
sparse_numeric
}
\description{
Multiply two sparse_numeric vectors elementwise
}
