% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Sparse Vector Multiplication}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{y}{A sparse_numeric vector}

\item{...}{Additional arguments}
}
\description{
Sparse Vector Multiplication

Multiply two sparse_numeric vectors elementwise
}
