% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{sparse_mult generic}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric vectors.}

\item{...}{Ignored, included for consistency with the generic.}
}
\value{
A sparse_numeric vector representing the element-wise product.
}
\description{
sparse_mult generic
}
