% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\alias{sparse_numeric-validity}
\title{Sparse Numeric Vector Class}
\description{
Sparse Numeric Vector Class

Validity method for sparse_numeric
}
\section{Slots}{

\describe{
\item{\code{pos}}{Integer vector of non-zero indices}

\item{\code{value}}{Numeric vector of non-zero values}

\item{\code{length}}{Integer giving total vector length}
}}

