% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseNumeric.R
\name{sparse_sub}
\alias{sparse_sub}
\title{Subtract two sparse_numeric vectors}
\usage{
sparse_sub(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric objects of same length}

\item{...}{unused}
}
\value{
sparse_numeric
}
\description{
Subtract two sparse_numeric vectors
}
