% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseNumeric.R
\name{standardize}
\alias{standardize}
\title{Standardize a sparse_numeric vector}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{sparse_numeric}

\item{...}{unused}
}
\value{
sparse_numeric
}
\description{
Each element is transformed to (x - mean(x)) / sd(x),
where mean and sd include implicit zeros.
}
