% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize Sparse Numeric Vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{...}{Additional arguments}
}
\description{
Standardize Sparse Numeric Vector

Standardize a sparse_numeric vector
}
