% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize an object}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{Object to standardize.}

\item{...}{Ignored.}
}
\value{
A new \linkS4class{sparse_numeric} object for sparse inputs.
}
\description{
Generic to create a standardized version of an object. For
\linkS4class{sparse_numeric} vectors, this centers and scales the
underlying dense vector (including zeros) so that it has mean 0 and
sample standard deviation 1.
}
