% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric vector.}

\item{...}{Ignored, included for consistency with the generic.}
}
\value{
A dense numeric vector of standardized values.
}
\description{
Subtract the mean and divide by the SD of the full vector.
}
